+++ app/src/main/java/woowacourse/shopping/data/db/cart/CartEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.local
+package woowacourse.shopping.data.db.cart
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
