+++ app/src/main/java/woowacourse/shopping/data/db/recent/RecentProductEntity.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data.local
+package woowacourse.shopping.data.db.recent
 
 import androidx.room.ColumnInfo
 import androidx.room.Entity
