+++ app/src/main/java/woowacourse/shopping/domain/repository/ProductRepository.kt
@@ -1,4 +1,6 @@
-package woowacourse.shopping.domain
+package woowacourse.shopping.domain.repository
+
+import woowacourse.shopping.domain.Product
 
 interface ProductRepository {
     fun load(
