+++ app/src/main/java/woowacourse/shopping/presentation/ui/ViewModelFactory.kt
@@ -2,9 +2,9 @@ package woowacourse.shopping.presentation.ui
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import woowacourse.shopping.data.local.AppDatabase
-import woowacourse.shopping.data.local.LocalCartRepository
-import woowacourse.shopping.data.local.LocalRecentProductRepository
+import woowacourse.shopping.data.db.AppDatabase
+import woowacourse.shopping.data.db.cart.RoomCartRepository
+import woowacourse.shopping.data.db.recent.LocalRecentProductRepository
 import woowacourse.shopping.data.remote.DummyProductRepository
 import woowacourse.shopping.presentation.ui.cart.CartViewModel
 import woowacourse.shopping.presentation.ui.detail.ProductDetailViewModel
@@ -18,7 +18,7 @@ class ViewModelFactory() : ViewModelProvider.Factory {
                 val cartDao = AppDatabase.instanceOrNull.cartDao()
                 ProductDetailViewModel(
                     DummyProductRepository(),
-                    LocalCartRepository(cartDao),
+                    RoomCartRepository(cartDao),
                     LocalRecentProductRepository(recentDao),
                 ) as T
             }
@@ -28,13 +28,13 @@ class ViewModelFactory() : ViewModelProvider.Factory {
                 val cartDao = AppDatabase.instanceOrNull.cartDao()
                 ShoppingViewModel(
                     recentRepository = LocalRecentProductRepository(recentDao),
-                    cartRepository = LocalCartRepository(cartDao),
+                    cartRepository = RoomCartRepository(cartDao),
                 ) as T
             }
 
             modelClass.isAssignableFrom(CartViewModel::class.java) -> {
                 val cartDao = AppDatabase.instanceOrNull.cartDao()
-                CartViewModel(cartRepository = LocalCartRepository(cartDao)) as T
+                CartViewModel(cartRepository = RoomCartRepository(cartDao)) as T
             }
 
             else -> {
