+++ app/src/main/java/woowacourse/shopping/presentation/ui/cart/adapter/CartAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.presentation.ui.cart
+package woowacourse.shopping.presentation.ui.cart.adapter
 
 import android.annotation.SuppressLint
 import android.view.LayoutInflater
@@ -6,6 +6,7 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemCartBinding
 import woowacourse.shopping.domain.ProductListItem
+import woowacourse.shopping.presentation.ui.cart.CartHandler
 
 class CartAdapter(
     private val cartHandler: CartHandler,
@@ -35,16 +36,5 @@ class CartAdapter(
 
     override fun getItemCount(): Int {
         return carts.size
-    }
-}
-
-class CartViewHolder(
-    private val binding: ItemCartBinding,
-    private val cartHandler: CartHandler,
-) :
-    RecyclerView.ViewHolder(binding.root) {
-    fun bind(item: ProductListItem.ShoppingProductItem) {
-        binding.cartHandler = cartHandler
-        binding.product = item
     }
 }
